pyproject.toml
[tool.poetry]
name = "perfi"
version = "0.0.1"
description = "Personal finances API"
authors = ["Thomas Davidoff <thomas.davidoff@gmail.com>"]
license = "MIT"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.11"
jinja2 = ">=3.1.5"
bcrypt = "^4.2.1"
greenlet = "^3.1.1"
pyyaml = "^6.0.2"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
pyjwt = "^2.10.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
pytest-cov = "^6.0.0"


[tool.poetry.scripts]
perfi = "cli:cli"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
alembic = "^1.14.0"
pre-commit = "^4.0.1"
psycopg2 = "^2.9.10"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
faker = "^33.1.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
addopts = "--capture=tee-sys -rP -v"
testpaths = ["tests"]
pythonpath = ["."]
